refactor: extract reusable show create formatters
