+++ ydb/core/sys_view/show_create/formatters/create_external_table_formatter.h
